Restoring the delta3d branch again.  Deleting dtDAL compatibility this time.

+++ branches/RB-2.8/inc/dtCore/namedunsignedlongintparameter.h
@@ -30,11 +30,11 @@ namespace dtCore
    /**
     * @class UnsignedLongIntMessageParameter
     */
-   class DT_CORE_EXPORT NamedUnsignedLongIntParameter: public NamedPODParameter<unsigned long>
+   class DT_CORE_EXPORT NamedUnsignedLongIntParameter: public NamedPODParameter<unsigned long long>
    {
       public:
          NamedUnsignedLongIntParameter(const dtUtil::RefString& name,
-            unsigned long defaultValue = 0, bool isList = false);
+            unsigned long long defaultValue = 0, bool isList = false);
 
       protected:
          virtual ~NamedUnsignedLongIntParameter();
